requirements.txt
python-pptx>=0.6.23
google-genai>=1.0.0
Pillow>=10.0.0
